+++ src/app/(services)/iphone/types/typeMakeIp.ts
@@ -1,6 +1,3 @@
-import MakeIp1 from "@/assets/img/iphone/makeIp1.png";
-import MakeIp2 from "@/assets/img/iphone/makeIp2.png";
-
 export type MakeIphone = {
     id: number;
     flex: string;
@@ -14,32 +11,3 @@ export type MakeIphone = {
     imgWidth?: number | string;
     imgHeight?: number | string;
 }
-
-export const arrMakeIphone: MakeIphone[] = [
-    {
-        id: 1,
-        flex: "col",
-        width: "full",
-        title: "iOS 16",
-        textCard: "Personal is powerful.",
-        padTop: 77,
-        padBottom: 0,
-        link: "",
-        img: MakeIp1,
-        imgWidth: "auto",
-        imgHeight: "auto",
-    },
-    {
-        id: 1,
-        flex: "col",
-        width: "auto",
-        title: "Switching to iPhone is super simple.",
-        textCard: "",
-        padTop: 61,
-        padBottom: 60,
-        link: "",
-        img: MakeIp2,
-        imgWidth: 675,
-        imgHeight: 357,
-    },
-];
